# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sahara AI Career & Wellness Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  sessionId: string;
  name?: string;
  interests?: string[];
  skills?: string[];
  education?: string;
  status?: string;
  weeklyLearningHours?: number;
  goals?: string;
  energyLevel?: number;
  stressLevel?: number;
  focusLevel?: number;
  targetCareer?: string;
  onboardingComplete: boolean;
  createdAt?: string;
}

export interface UpdateProfileRequest {
  name?: string;
  interests?: string[];
  skills?: string[];
  education?: string;
  status?: string;
  weeklyLearningHours?: number;
  goals?: string;
  energyLevel?: number;
  stressLevel?: number;
  focusLevel?: number;
  targetCareer?: string;
}

export interface OnboardingData {
  name?: string;
  interests: string[];
  skills: string[];
  education: string;
  status: string;
  weeklyLearningHours: number;
  goals: string;
  energyLevel?: number;
  stressLevel?: number;
  focusLevel?: number;
}

export interface CareerRecommendation {
  id: string;
  title: string;
  whyItFits: string;
  difficulty: string;
  requiredSkills: string[];
  marketDemand: string;
  salaryRange: string;
  matchScore: number;
}

export interface CareerRecommendationsResponse {
  recommendations: CareerRecommendation[];
}

export interface RealitySimulatorResponse {
  careerId: string;
  careerTitle: string;
  dailyTasks: string[];
  workEnvironment: string;
  effortLevel: string;
  timeline: string;
  challenges: string[];
  typicalDay: string;
  salaryJourney: string;
  burnoutRisk: string;
}

export interface CareerProbabilityResponse {
  careerId: string;
  careerTitle: string;
  successProbability: number;
  requiredEffort: string;
  suggestedLearningTime: string;
  keyFactors: string[];
  improvementTips: string[];
}

export interface Skill {
  id: string;
  name: string;
  currentLevel: number;
  targetLevel: number;
  category: string;
  resources: string[];
}

export interface Milestone {
  id: string;
  title: string;
  description: string;
  completed: boolean;
  targetDate?: string;
}

export interface SkillGapsResponse {
  targetCareer: string;
  overallProgress: number;
  skills: Skill[];
  milestones: Milestone[];
  missingSkills: string[];
}

export interface UpdateSkillProgressRequest {
  skillId: string;
  currentLevel: number;
}

export interface WellnessData {
  stressLevel: number;
  energyLevel: number;
  focusLevel: number;
  moodScore: number;
  lastCheckin?: string;
  recommendations?: string[];
}

export interface WellnessCheckinRequest {
  stressLevel: number;
  energyLevel: number;
  focusLevel: number;
  moodScore: number;
  notes?: string;
}

export interface WellnessExercise {
  id: string;
  title: string;
  description: string;
  duration: string;
  type: string;
}

export interface WellnessCheckinResponse {
  wellness: WellnessData;
  recommendations: string[];
  exercises: WellnessExercise[];
}

export interface MoodEntry {
  date: string;
  moodScore: number;
  stressLevel: number;
  energyLevel: number;
}

export interface MoodHistoryResponse {
  entries: MoodEntry[];
}

export interface Task {
  id: string;
  title: string;
  description: string;
  category: string;
  priority: string;
  completed: boolean;
  dueDate?: string;
  estimatedMinutes: number;
  isWellnessBreak: boolean;
}

export interface WeeklyPlanResponse {
  weekStart: string;
  weekEnd: string;
  completionPercentage: number;
  tasks: Task[];
  focusTip: string;
  adaptiveNote: string;
}

export interface ChatMessageRequest {
  message: string;
  sessionId?: string;
}

export interface ChatMessage {
  id: string;
  role: string;
  content: string;
  timestamp: string;
}

export interface ChatMessageResponse {
  message: ChatMessage;
  suggestions: string[];
}

export interface ChatHistoryResponse {
  messages: ChatMessage[];
}

export interface JobRole {
  title: string;
  demandLevel: string;
  avgSalary: string;
  growthRate: string;
  topSkills: string[];
}

export interface JobMarketResponse {
  trending: JobRole[];
  insights: string[];
  topIndustries: string[];
}

export interface DashboardResponse {
  profile?: UserProfile;
  careerProgress: number;
  weeklyTasksCompleted: number;
  weeklyTasksTotal: number;
  wellness: WellnessData;
  upcomingTasks: Task[];
  streakDays: number;
}