# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sahara AI Career & Wellness Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /tasks/{taskId}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}/uncomplete:
    post:
      operationId: uncompleteTask
      tags: [tasks]
      summary: Mark a task as incomplete
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        name:
          type: string
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        education:
          type: string
        status:
          type: string
        weeklyLearningHours:
          type: number
        goals:
          type: string
        energyLevel:
          type: number
        stressLevel:
          type: number
        focusLevel:
          type: number
        targetCareer:
          type: string
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - onboardingComplete
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        education:
          type: string
        status:
          type: string
        weeklyLearningHours:
          type: number
        goals:
          type: string
        energyLevel:
          type: number
        stressLevel:
          type: number
        focusLevel:
          type: number
        targetCareer:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
        completed:
          type: boolean
        dueDate:
          type: string
        estimatedMinutes:
          type: number
        isWellnessBreak:
          type: boolean
      required:
        - id
        - title
        - description
        - category
        - priority
        - completed
        - estimatedMinutes
        - isWellnessBreak
...